% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_t1_mapa_raster.R
\name{addBasemapTiles}
\alias{addBasemapTiles}
\title{Agrega un mapa base con relieve de Esri}
\usage{
addBasemapTiles(map)
}
\arguments{
\item{map}{Mapa Leaflet al que se le añadirá el mapa base con relieve.}
}
\value{
El mapa Leaflet con el mapa base añadido.
}
\description{
Agrega un mapa base de relieve (Esri World Imagery) al mapa Leaflet,
utilizado por defecto en el mapa de la provincia de Buenos Aires.
}
