% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_t1_mapa_raster.R
\name{addPolygonsLayer}
\alias{addPolygonsLayer}
\title{Agrega polígonos de un conjunto de datos al mapa Leaflet}
\usage{
addPolygonsLayer(mapa, data, fillopacity_poly)
}
\arguments{
\item{mapa}{Mapa Leaflet al que se le añadirán los polígonos.}

\item{data}{Conjunto de datos de clase \code{sf} que contiene la geometría de los polígonos.}

\item{fillopacity_poly}{Valor numérico entre 0 y 1 que especifica la opacidad de relleno de los polígonos.}
}
\value{
El mapa Leaflet con la capa de polígonos añadida.
}
\description{
Esta función agrega una capa de polígonos representando los partidos de la provincia de Buenos Aires
al mapa Leaflet. Cada polígono tiene un estilo de borde y relleno configurable.
}
