% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_t1_mapa_raster.R
\name{addRasterLegend}
\alias{addRasterLegend}
\title{Agregar imagen raster y leyenda a un mapa Leaflet}
\usage{
addRasterLegend(mapa, imagen, opacidad, paleta, etiquetas)
}
\arguments{
\item{mapa}{Objeto \code{leaflet}. Mapa al que se le añadirá la imagen raster y la leyenda.}

\item{imagen}{Objeto \code{raster} o \code{SpatRaster} (por ejemplo, de \code{terra} o \code{raster}) que representa la capa raster a visualizar.}

\item{opacidad}{\code{numeric}. Valor de opacidad para la capa raster (0 a 1).}

\item{paleta}{Objeto \code{colorBin} o \code{colorNumeric} generado con \code{leaflet::colorBin} o \code{leaflet::colorNumeric} para definir la escala de colores.}

\item{etiquetas}{\code{character}. Vector de etiquetas que se usará para mostrar los intervalos en la leyenda.}
}
\value{
Un objeto \code{leaflet} con la capa raster y la leyenda añadidas.
}
\description{
Esta función agrega una imagen raster a un mapa interactivo creado con \code{leaflet} y añade una leyenda personalizada que muestra el porcentaje de cambio
de los valores del raster. La leyenda se construye a partir de una paleta de colores binaria o numérica y se ubica en la esquina superior derecha.
También se asegura de eliminar capas o leyendas previas para evitar duplicaciones.
}
