% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_amba.R
\docType{data}
\name{amba_caba}
\alias{amba_caba}
\title{Nombres del AMBA con Capital Federal}
\format{
Vector de caracteres de longitud 34.
}
\usage{
amba_caba
}
\description{
Vector de caracteres con los nombres seleccionados de partidos y la "Capital Federal"
que representa a la Ciudad Autónoma de Buenos Aires (CABA) en el Área Metropolitana de Buenos Aires.
}
\examples{
data(amba_caba)
head(amba_caba)
}
\keyword{datasets}
