% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_amba.R
\docType{data}
\name{amba_reducido_names}
\alias{amba_reducido_names}
\title{Nombres reducidos del AMBA}
\format{
Vector de caracteres de longitud 38.
}
\usage{
amba_reducido_names
}
\description{
Vector de caracteres con los nombres seleccionados de partidos y comunas
del Área Metropolitana de Buenos Aires (AMBA).
Incluye 15 comunas de CABA con el prefijo "Comuna".
}
\examples{
data(amba_reducido_names)
head(amba_reducido_names)
}
\keyword{datasets}
