% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_t1_barra_informativa_inferior.R
\name{BarraInferior_Server}
\alias{BarraInferior_Server}
\title{Logica del Servidor para la Barra Informativa Inferior}
\usage{
BarraInferior_Server(id, boton, fecha, area, momento, porcentaje)
}
\arguments{
\item{id}{Cadena de caracteres usada para identificar el espacio de nombres del modulo.}

\item{boton}{Expresion reactiva para el boton de actualizacion (por ejemplo, \code{input$update_map}).}

\item{fecha}{Expresion reactiva que retorna la fecha seleccionada como cadena.}

\item{area}{Expresion reactiva que indica el area seleccionada (por ejemplo, "amba").}

\item{momento}{Expresion reactiva que indica el momento del dia (por ejemplo, "manana", "tarde").}

\item{porcentaje}{Expresion reactiva que indica el tipo de dato (por ejemplo, "pc" para prepandemia).}
}
\value{
No retorna valor. Se llama por efectos secundarios para actualizar dinamicamente el texto en la UI.
}
\description{
Logica del lado servidor para manejar el texto informativo mostrado en la
barra inferior. Reacciona a la interaccion del usuario y cambios en la entrada
para actualizar el mensaje.
}
