% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_t1_barra_informativa_inferior.R
\name{BarraInferior_UI}
\alias{BarraInferior_UI}
\title{Componente de UI para Barra Informativa Inferior}
\usage{
BarraInferior_UI(id)
}
\arguments{
\item{id}{Cadena de caracteres usada para el espacio de nombres de los elementos UI.}
}
\value{
Un elemento UI (tagList) que contiene un div con salida de texto estilizada.
}
\description{
Crea un elemento de interfaz de usuario que muestra texto informativo dinamico
en la parte inferior de la aplicacion Shiny.
}
