% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_base_raster.R
\docType{data}
\name{base_raster}
\alias{base_raster}
\title{Información de archivos raster usados en el proyecto}
\format{
Un data frame con \code{n} filas y 7 columnas:
\describe{
\item{filename}{Nombre del archivo raster}
\item{file_info}{Información adicional sobre el archivo (puede incluir tamaño, formato, etc.)}
\item{locacion}{Ubicación geográfica o área asociada al raster}
\item{tipo_de_raster}{Tipo de raster (por ejemplo, 'pc', '7dpc', etc.)}
\item{fecha}{Fecha asociada al raster}
\item{hora}{Hora asociada al raster}
\item{momento}{Momento o periodo relacionado con el raster (por ejemplo, pre-pandemia, semanal, etc.)}
}
}
\source{
Datos generados y procesados para la aplicación GeoCovid
}
\usage{
base_raster
}
\description{
Conjunto de datos que contiene metadata sobre los archivos raster empleados en el análisis,
incluyendo información del archivo, ubicación geografica, tipo de raster, fecha, hora y momento.
}
\details{
Para una explicación más detallada sobre el origen y procesamiento de estos datos,
consulte el sitio web del proyecto GeoCovid Buenos Aires:
\url{https://fernandez-lab-wsu.github.io/geocovid_bsas/} y el código en
el archivo \code{tools/helper_datasets.R} incluido en este paquete.
}
\keyword{datasets}
