% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_t1_boton_ayuda.R
\name{BotonAyuda_UI}
\alias{BotonAyuda_UI}
\title{UI Component for Help Button with Tooltip}
\usage{
BotonAyuda_UI(id)
}
\arguments{
\item{id}{Character string used to namespace the UI elements.}
}
\value{
A UI element (tagList) containing a help button and a hidden tooltip.
}
\description{
Creates a help button with an associated tooltip that appears on hover.
This module is intended to be used within a Shiny UI.
}
