% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_centroides.R
\docType{data}
\name{centroides_mapa}
\alias{centroides_mapa}
\title{Coordenadas de los centroides de los partidos de Buenos Aires}
\format{
Un data frame con 136 filas y 3 columnas:
\describe{
\item{partido}{Nombre del partido.}
\item{X}{Longitud del centroide del partido (coordenada geográfica).}
\item{Y}{Latitud del centroide del partido (coordenada geográfica).}
}
}
\source{
Extraído de los mapas vectoriales oficiales.
}
\usage{
data(centroides_mapa)
}
\description{
Un conjunto de datos con las coordenadas geográficas aproximadas (longitud y latitud) de los centroides de los partidos de la Provincia de Buenos Aires, Argentina.
}
\keyword{datasets}
