% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_t2t3_dygraph_casos_covid.R
\name{convertir_xts_serie}
\alias{convertir_xts_serie}
\title{Convertir datos de casos COVID-19 a formato xts}
\usage{
convertir_xts_serie(serie, nombre_serie)
}
\arguments{
\item{serie}{Nombre de la provincia (por ejemplo, \code{"Buenos Aires"} o \code{"CABA"}), usado para filtrar los datos.}

\item{nombre_serie}{Nombre deseado para la columna de valores (por ejemplo, \code{"casos_bsas"}).}
}
\value{
Un objeto de clase \code{xts} con las fechas como índice y una columna con los casos, renombrada como \code{nombre_serie}.
}
\description{
Esta función filtra los datos de \code{geocovidapp::casos_covid_dpto} para una provincia específica,
renombra la columna de casos (\code{n}) con un nombre personalizado, y convierte el resultado a un objeto \code{xts}.
}
\examples{
\dontrun{
convertir_xts_serie("CABA", "casos_caba")
}

}
