% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_t2t3_grafico_casos_covid.R
\name{Dygraph_Server}
\alias{Dygraph_Server}
\title{Servidor: Gráficos de casos de COVID por provincia y departamento}
\usage{
Dygraph_Server(id, area, partido)
}
\arguments{
\item{id}{Módulo Shiny (string) usado como identificador del namespace.}

\item{area}{Reactive que indica si el raster corresponde a Buenos Aires provincia o a AMBA.}

\item{partido}{Reactive que devuelve el nombre del partido de la provincia de Buenos Aires,
seleccionado en otro módulo.}
}
\value{
Una lista con tres reactives:
\describe{
\item{casos_covid}{La fecha seleccionada por el usuario en el gráfico de COVID-19.}
\item{grafico_casos_prov}{Reactive con el gráfico de casos por provincia o CABA.}
\item{grafico_casos_dpto}{Reactive con el gráfico de casos por departamento o comuna.}
}
}
\description{
Servidor: Gráficos de casos de COVID por provincia y departamento
}
