% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_t2t3_grafico_casos_covid.R
\name{Dygraph_UI}
\alias{Dygraph_UI}
\title{Interfaz de usuario: Gráficos de casos de covid por provincia y departamento}
\usage{
Dygraph_UI(id)
}
\arguments{
\item{id}{Module name}
}
\value{
Dos gráficos de casos de COVID-19 en el tiempo.
}
\description{
Interfaz de usuario: Gráficos de casos de covid por provincia y departamento
}
