% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_t3_extrae_px_data.R
\name{extrae_px_data}
\alias{extrae_px_data}
\title{Extraer datos de px desde la base de datos}
\usage{
extrae_px_data(partido, tipo_tab, fecha, mapa, con = NULL, demo = FALSE)
}
\arguments{
\item{partido}{Character. Nombre del partido (necesario si \code{mapa = FALSE}).}

\item{tipo_tab}{Character. Tipo de raster a consultar.}

\item{fecha}{Character. Fecha en formato 'YYYY-MM-DD' (necesaria si \code{mapa = TRUE}).}

\item{mapa}{Logical. Indica si se consulta la versión para mapa (\code{TRUE}) o tabla (\code{FALSE}).}

\item{con}{Conexión a la base de datos (objeto DBI).}

\item{demo}{Logical. Si es \code{TRUE}, carga datos ficticios en lugar de consultar la base.}
}
\value{
Un objeto espacial (sf) con los datos consultados.
}
\description{
Esta función obtiene datos espaciales desde la base de datos según los parámetros
especificados, con la opción de cargar datos de prueba en modo demo.
}
\details{
\itemize{
\item Si \code{mapa = FALSE}, se requiere el parámetro \code{partido} y se consulta la tabla 'px_tab3_promedio'.
\item Si \code{mapa = TRUE}, se requiere el parámetro \code{fecha} y se consulta la tabla con datos promedio por fecha.
\item En modo \code{demo = TRUE} se retornan datos ficticios para pruebas.
}
}
