% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_t1_fecha_momento.R
\name{FechaMomento_UI}
\alias{FechaMomento_UI}
\title{Elementos de interfaz de usuario de la barra flotante del tab de Mapa BsAs}
\usage{
FechaMomento_UI(id)
}
\arguments{
\item{id}{Nombre del modulo.}
}
\value{
Elementos de interfaz de usuario para la barra flotante en el tab Mapa Buenos Aires.
}
\description{
Esta funcion crea los elementos de la interfaz de usuario para la barra flotante
en el tab de Mapa Buenos Aires, que permite al usuario seleccionar opciones como el
mapa base, area de interes, fecha, momento del dia y opciones de transparencia.
}
