% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_conversion_fecha.R
\name{formatted_date}
\alias{formatted_date}
\title{Convierte el formato de la fecha de \%Y-\%m-\%dT\%H:\%M:\%S a "\%Y-\%m-\%d"}
\usage{
formatted_date(fecha)
}
\arguments{
\item{fecha}{Fecha}
}
\value{
Fecha en formato "\%Y-\%m-\%d"
}
\description{
Convierte el formato de la fecha de \%Y-\%m-\%dT\%H:\%M:\%S a "\%Y-\%m-\%d"
}
\examples{

formatted_date("2020-05-10T07:00:00.000Z")

}
