% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{geocovidapp_init}
\alias{geocovidapp_init}
\title{Inicializador de la aplicación GeoCovidApp}
\usage{
geocovidapp_init()
}
\value{
Un objeto de tipo \code{list} con objetos inicializados, como por ejemplo \code{pool} o configuración.
}
\description{
Esta función se utiliza para inicializar la aplicación GeoCovidApp.
Es recomendable llamarla al comienzo de la función \code{run_app()} para realizar tareas
de configuración necesarias antes de que se ejecute la interfaz de usuario y el servidor.
}
\details{
Por ejemplo, esta función puede:
\itemize{
\item Leer archivos de configuración con el paquete \code{config}.
\item Establecer conexiones a bases de datos usando \code{pool}.
\item Definir opciones globales o variables necesarias durante la ejecución de la app.
}
}
