% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_t2_reporte.R
\name{ggplot_casos_covid_doble}
\alias{ggplot_casos_covid_doble}
\title{Grafico individual y comparativo de casos COVID}
\usage{
ggplot_casos_covid_doble(partido, fecha)
}
\arguments{
\item{partido}{Nombre del partido o comuna (ej. "La Matanza" o "Comuna 1")}

\item{fecha}{Fecha a marcar con una linea vertical (formato "YYYY-MM-DD")}
}
\value{
Una lista con dos ggplots: \code{grafico_partido} y \code{grafico_comparativo}
}
\description{
Grafico individual y comparativo de casos COVID
}
