% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_t2_histograma_raster.R
\name{HistogramaRaster_UI}
\alias{HistogramaRaster_UI}
\title{IU: Histograma de la movilidad ciudadana por departamento}
\usage{
HistogramaRaster_UI(id)
}
\arguments{
\item{id}{Module name}
}
\value{
Histograma de la movilidad ciudadana por partido, respetando la
escala de colores del raster.
}
\description{
IU: Histograma de la movilidad ciudadana por departamento
}
