% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_demo_data.R
\name{load_fake_px_data}
\alias{load_fake_px_data}
\title{Genera un conjunto de datos espaciales falsos tipo \code{px_promedio}}
\usage{
load_fake_px_data(partido, tipo_tab)
}
\arguments{
\item{partido}{Nombre del partido (municipio) a simular. Se utiliza como valor
de la columna \code{partido}.}

\item{tipo_tab}{Tipo de raster o categoría de análisis, que se guarda en la
columna \code{tipo_de_raster}.}
}
\value{
Un objeto \code{sf::sf} con 4 polígonos simples y las siguientes columnas:
\itemize{
\item \code{fecha}: Fechas consecutivas simuladas.
\item \code{locacion}: Siempre \code{"baires"} en esta simulación.
\item \code{tipo_de_raster}: Tipo de raster indicado como argumento.
\item \code{partido}: Nombre del partido pasado como argumento.
\item \code{manana_8}, \code{tarde_16}, \code{noche_0}: Valores aleatorios entre 0 y 7.
\item \code{px_mean_dianoche}: Promedio entre \code{manana_8} y \code{tarde_16}.
\item \code{criterio}: Clasificación basada en \code{px_mean_dianoche}.
\item \code{criterio_noche}: Clasificación basada en \code{noche_0}.
}
}
\description{
Esta función genera un objeto \code{sf} que simula la estructura de datos
provenientes de una base de datos \code{px_promedio}, incluyendo variables horarias
y criterios calculados a partir de los valores simulados. Útil para pruebas
en modo demo.
}
\details{
Los criterios (\code{criterio}, \code{criterio_noche}) se asignan según los siguientes rangos:
\itemize{
\item Menor a 2.5: \code{"5 - 1"}
\item Entre 2.5 y 5: \code{"10 - 1"}
\item Mayor o igual a 5: \code{"20 - 1"}
}
}
