% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_demo_data.R
\name{load_fake_raster}
\alias{load_fake_raster}
\title{Genera un raster falso para pruebas en modo demo}
\usage{
load_fake_raster(area)
}
\arguments{
\item{area}{Área geográfica para la que se quiere simular el raster.
Puede ser \code{"baires"} (provincia de Buenos Aires), \code{"amba"} (área metropolitana de Buenos Aires),
o cualquier otro valor para una región genérica por defecto.}
}
\value{
Un objeto \code{terra::SpatRaster} con valores numéricos aleatorios entre -50 y 50,
sistema de referencia geográfico WGS84 (EPSG:4326), y extensión definida según el área.
}
\description{
Esta función genera un objeto \code{SpatRaster} con valores aleatorios entre -50 y 50,
útil para pruebas en modo demo sin conexión a bases de datos reales.
La extensión y resolución del raster varían según el área especificada.
}
