% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_t2_reporte.R
\name{mapa_reporte}
\alias{mapa_reporte}
\title{Generar Mapa para Reporte}
\usage{
mapa_reporte(part, imagen)
}
\arguments{
\item{part}{Nombre del partido a visualizar. Debe coincidir con la columna \code{partido} del objeto \code{geocovidapp::bsas_comunas}.}

\item{imagen}{Objeto raster de clase \code{SpatRaster} (del paquete\code{terra}) que contiene los datos a mapear.}
}
\value{
Un objeto \code{ggplot} que muestra el mapa tematico del area seleccionada.
}
\description{
Crea un mapa con un raster reclasificado y coloreado por rangos de porcentaje de cambio.
El area mostrada corresponde al partido especificado. Se utiliza \code{ggplot2} y \code{tidyterra} para la visualizacion.
}
