% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_t1_mapa_baires.R
\name{MapaBaires_UI}
\alias{MapaBaires_UI}
\title{IU: Mapa raster de la provincia de Buenos Aires}
\usage{
MapaBaires_UI(id)
}
\arguments{
\item{id}{Nombre del módulo. Este parámetro es necesario para la correcta
integración del módulo UI dentro de la aplicación Shiny.}
}
\value{
Un objeto UI de Shiny que contiene un mapa interactivo de la provincia de Buenos Aires,
renderizado con \code{leaflet}. El mapa es capaz de manejar eventos de interacción como zoom y desplazamiento.
}
\description{
Esta función crea la interfaz de usuario (UI) para el mapa interactivo
de la provincia de Buenos Aires, que se encuentra en el primer tab de la aplicación.
Utiliza el paquete \code{leaflet} para mostrar un mapa interactivo con un alto grado de interactividad,
permitiendo a los usuarios visualizar y explorar los datos geoespaciales de la región.
}
