% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_t3_mapa_covid_departamentos.R
\name{MapaCovidDepartamentos_Server}
\alias{MapaCovidDepartamentos_Server}
\title{Server: Modulo mapa casos COVID con promedio por departamento de Bs. As.}
\usage{
MapaCovidDepartamentos_Server(id, pool)
}
\arguments{
\item{id}{\code{character}. Identificador del modulo Shiny.}

\item{pool}{Conexion a base de datos.}
}
\value{
Una lista de reactivos con el raster seleccionado (\code{mapa_partido}) y el nivel de zoom (\code{zoom_mapa_partido}).
}
\description{
Logica del servidor para el modulo que muestra un mapa de Buenos Aires con un
promedio de los valores de raster para ese dia y los casos de COVID-19 por
departamento como burbujas.
Este modulo corresponde al tab 3, "Por partido", de la aplicacion GeoCovid.
}
