% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_t3_mapa_covid_elecciones_input.R
\name{MapaCovidElecciones_Server}
\alias{MapaCovidElecciones_Server}
\title{Servidor: Selector de fecha}
\usage{
MapaCovidElecciones_Server(id)
}
\arguments{
\item{id}{Module name}
}
\value{
Fecha seleccionada por el usuario
}
\description{
Servidor: Selector de fecha
}
