% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_t2_mapa_partido.R
\name{MapaPartido_Server}
\alias{MapaPartido_Server}
\title{Server: Módulo de mapa raster por partido}
\usage{
MapaPartido_Server(
  id,
  act_mapas,
  imagen,
  area,
  fecha,
  tipo_de_raster,
  opacidad,
  par_d_bsas,
  momento_dia
)
}
\arguments{
\item{id}{\code{character}. Identificador del módulo Shiny.}

\item{act_mapas}{\code{reactive}. Trigger reactivo para actualizar el mapa.}

\item{imagen}{\code{reactive}. Lista de rasters.}

\item{area}{\code{character}. Indica si el raster corresponde a Buenos Aires o AMBA.}

\item{fecha}{\code{Date} o \code{reactive}. Fecha seleccionada.}

\item{tipo_de_raster}{\code{character}. Define si el raster es de cambio porcentual prepandemia (\code{pc}) o semanal (\verb{7dpc}).}

\item{opacidad}{\code{numeric}. Valor de opacidad del raster.}

\item{par_d_bsas}{\code{reactive}. Partido seleccionado (provincia de Buenos Aires).}

\item{momento_dia}{\code{numeric}. Hora del día representada: 0, 8 o 16 horas.}
}
\value{
Una lista de reactivos con el raster seleccionado (\code{mapa_partido}) y el nivel de zoom (\code{zoom_mapa_partido}).
}
\description{
Lógica del servidor para el módulo que muestra un mapa raster por partido.
Este módulo corresponde a la pestaña 2, "Por partido", de la aplicación GeoCovid.
}
