% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_t2_mapa_partido.R
\name{MapaPartido_UI}
\alias{MapaPartido_UI}
\title{IU: Mapa raster por partido}
\usage{
MapaPartido_UI(id)
}
\arguments{
\item{id}{Module name}
}
\value{
Mapa leaflet en la IU
}
\description{
Este mapa se encuentra en el tab 2, "por partido" de GeoCovid app
}
