% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_t2_partidos_input.R
\name{Partidos_Server}
\alias{Partidos_Server}
\title{Servidor: Seleccion de departamentos de provincia de Buenos Aires con IU
dinámica.}
\usage{
Partidos_Server(id)
}
\arguments{
\item{id}{Module name}
}
\value{
El area y el partido seleccionado.
}
\description{
Servidor: Seleccion de departamentos de provincia de Buenos Aires con IU
dinámica.
}
