% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_t2_partidos_input.R
\name{Partidos_UI}
\alias{Partidos_UI}
\title{IU: Seleccion de departamentos de provincia de Buenos Aires con IU
dinámica.}
\usage{
Partidos_UI(id)
}
\arguments{
\item{id}{Module name}
}
\value{
Elementos de la IU para que se seleccione partidos de AMBA o de
BsAs de forma condicional.
}
\description{
IU: Seleccion de departamentos de provincia de Buenos Aires con IU
dinámica.
}
