% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_pixel_promedi_DEL.R
\name{pixel_loader}
\alias{pixel_loader}
\title{Cargar pixeles raster desde base de datos}
\usage{
pixel_loader(partido, tipo_de_raster)
}
\arguments{
\item{partido}{\code{character}. Nombre del partido para filtrar.}

\item{tipo_de_raster}{\code{character}. Tipo de raster a consultar (por ejemplo, \code{pc} o \verb{7dpc}).}
}
\value{
Un objeto \code{sf} con los pixeles raster filtrados.
}
\description{
Conecta a una base de datos PostgreSQL/PostGIS y extrae pixeles raster
según el partido y el tipo de raster especificados.
}
\details{
La conexión se configura usando los parámetros definidos en el archivo \code{config.yml}.
Utiliza \code{pool} para gestionar la conexión y \code{sf::st_read} para leer datos espaciales.
}
