% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_raster_valido.R
\name{raster_valido}
\alias{raster_valido}
\title{Verificar si la combinación de parámetros existe en el dataframe \code{base_raster}}
\usage{
raster_valido(area, porcentaje, fecha, momento, base_raster)
}
\arguments{
\item{area}{Una cadena de texto que representa el área seleccionada por el usuario (por ejemplo, "baires" o "amba").}

\item{porcentaje}{Una cadena de texto que representa el tipo de cambio porcentual seleccionado por el usuario (por ejemplo, "pc" o "7dpc").}

\item{fecha}{Un objeto de fecha que representa la fecha seleccionada por el usuario.}

\item{momento}{Una cadena de texto que representa el momento del día seleccionado por el usuario (por ejemplo, "mañana", "tarde" o "noche").}

\item{base_raster}{Un dataframe que contiene los rasters con las características de locación, tipo de raster, fecha y momento del día.}
}
\value{
\code{TRUE} si existe una fila en \code{base_raster} que cumpla con todas las condiciones especificadas; de lo contrario, \code{FALSE}.
}
\description{
Esta función toma los valores de área, tipo de raster, fecha y momento del día
seleccionados por el usuario y verifica si existe una fila en el dataframe \code{base_raster}
que coincida con todos estos valores.
Es una funcion defensiva, pensada para tratar de encontrar los casos
cuando no hay rasters y evitar que la aplicacion se rompa
}
