% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_raster_loader.R
\name{rasterLoader}
\alias{rasterLoader}
\title{Carga de raster desde la base de datos o raster falso en modo demo}
\usage{
rasterLoader(pool, raster_data, area, demo = FALSE)
}
\arguments{
\item{pool}{Conexion al pool de base de datos (puede ser \code{NULL} si \code{demo = TRUE}).}

\item{raster_data}{Un dataframe que contiene la informacion del raster, incluyendo el nombre del archivo (\code{filename}).}

\item{area}{Cadena de texto que indica el área de interés. Puede ser \code{"amba"} o \code{"baires"}.}

\item{demo}{Logico, si es \code{TRUE} carga un raster falso para modo demo. Por defecto \code{FALSE}.}
}
\value{
Un objeto \code{rast} del paquete \code{terra}, que representa el raster cargado (real o simulado).
}
\description{
Esta funcion carga un raster desde la base de datos segun el area especificada
("amba" o "baires"). Utiliza una consulta SQL para recuperar el raster como un archivo binario
y lo guarda temporalmente como un archivo \code{.tif}, que luego es cargado como un objeto raster
utilizando el paquete \code{terra}.
Si el argumento \code{demo} es \code{TRUE}, devuelve un raster simulado sin necesidad de conexion a la base.
}
