% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_t2_reporte.R
\name{ReporteServer}
\alias{ReporteServer}
\title{Servidor: reporte personalizado por partido}
\usage{
ReporteServer(
  id,
  partido,
  fecha,
  area,
  act_mapas,
  tipo_de_raster,
  opacidad,
  mapa_partido_manana,
  mapa_partido_tarde,
  mapa_partido_noche
)
}
\arguments{
\item{id}{Nombre del módulo. Necesario para su uso con \code{moduleServer}.}

\item{partido}{Reactive. Nombre del partido seleccionado.}

\item{fecha}{Reactive. Fecha seleccionada por el usuario.}

\item{area}{Reactive. Área geográfica del raster: \code{"baires"} o \code{"amba"}.}

\item{act_mapas}{Reactive. Evento que activa la actualización del mapa. Se asegura asi que el reporte este actualizado con la ultima eleccion del usuario.}

\item{tipo_de_raster}{Reactive. Tipo de comparación del raster: \code{"pc"} (prepandemia) o \code{"7dpc"} (últimos 7 días).}

\item{opacidad}{Reactive. Valor numérico entre 0 y 1 que define la opacidad del raster.}

\item{mapa_partido_manana}{Reactive. Raster correspondiente al momento de la mañana.}

\item{mapa_partido_tarde}{Reactive. Raster correspondiente al momento de la tarde.}

\item{mapa_partido_noche}{Reactive. Raster correspondiente al momento de la noche.}
}
\value{
Este módulo no retorna valores al servidor de forma explícita, pero registra un \code{downloadHandler}
que permite descargar un reporte en formato Word (\code{.docx}) con los parámetros seleccionados.
}
\description{
Este módulo corresponde al Tab 2: "Por partido" de la aplicación GeoCovid. Permite generar
un reporte descargable en formato Word (\code{.docx}) basado en la selección de un partido de
la provincia de Buenos Aires, una fecha, y los datos de movilidad ciudadana.
}
