% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_t2_reporte.R
\name{ReporteUI}
\alias{ReporteUI}
\title{IU: reporte}
\usage{
ReporteUI(id)
}
\arguments{
\item{id}{Module name}
}
\value{
Botón de descarga del reporte.
}
\description{
Este módulo se encuentra en el tab 2: 'Por partido' de GeoCovid app
}
