% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Ejecuta la aplicación Shiny GeoCovidApp}
\usage{
run_app(demo = FALSE)
}
\arguments{
\item{demo}{Lógico. Si es \code{TRUE}, activa el modo demo (por defecto \code{FALSE}).}
}
\value{
Lanza una instancia de la app Shiny y no devuelve ningún valor.
}
\description{
Esta función inicializa y ejecuta la aplicación principal de GeoCovidApp.
Se encarga de:
}
\details{
\itemize{
\item Establecer la conexión a la base de datos mediante un pool.
\item Cargar los datasets internos del paquete (e.g., \code{base_raster}, \code{bsas}, etc.).
\item Llamar a \code{geocovidapp_init()} para configurar recursos adicionales.
\item Ejecutar la aplicación Shiny con \code{ui()} y \code{server()}.
}
}
