% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui}
\alias{ui}
\title{Interfaz de usuario de la aplicacion GeoCovid}
\usage{
ui()
}
\value{
Un objeto UI para la Shiny app.
}
\description{
Genera la interfaz grafica principal (UI) de la aplicacion Shiny GeoCovid.
Incluye:
\itemize{
\item Un tema personalizado con colores y fuentes especificas usando \code{bslib}.
\item Pestanas para movilidad en Buenos Aires, analisis por partido, panorama general, y una seccion sobre el proyecto.
\item Paneles flotantes, controles deslizantes, mapas, graficos interactivos y contenido estatico.
\item Inclusion de CSS personalizado y uso de \code{shinyjs} para funcionalidades JavaScript.
}
}
